Arg for begin number; special column char pipe
I use `ruler` in my `tmux` status line to show 80 columns. I also use a
personal log with a timestamp at the front.

I put the status line clock on the left for easy comparison, but needed to add
a begin line for ruler to have the correct position information.

There remains a bug with the special characters surrounding column 80 where it
isn't adjusting for the new start number.

To make the bug easier to see I changed the special character to a vertical bar
(or pipe) which seemed better than an apostraphe.